+++ src/settingsmanager.cpp
@@ -10,7 +10,7 @@
 
 // Injected by qmake from the spec file version
 #ifndef APP_VERSION
-#define APP_VERSION "2.2.0"
+#define APP_VERSION "2.2.1"
 #endif
 
 namespace {